refactor: move WheelMotor implementation to separate file
